+++ src/sll/include/sll/types.h
@@ -941,10 +941,18 @@ typedef struct __SLL_VAR_ARG_LIST_DATA_STRUCT{
 
 
 
+typedef struct __SLL_VAR_ARG_LIST_DATA_ARRAY{
+	void*const* p;
+	sll_arg_count_t l;
+} sll_var_arg_list_data_array_t;
+
+
+
 typedef union __SLL_VAR_ARG_LIST_DATA{
 	va_list* c;
 	sll_var_arg_list_data_sll_t sll;
 	sll_var_arg_list_data_struct_t s;
+	sll_var_arg_list_data_array_t a;
 } sll_var_arg_list_data_t;